requirements.txt
fastapi
uvicorn
python-multipart
pypdf
openai
python-dotenv
jinja2
google-genai
